Update minigame success criteria
Changed the 'maxTries' variable to 'requiredClicks' in the Minigame class. This change is implemented to clarify that the game's success depends on getting the required number of correct clicks, rather than just trying a max number of times. This adjustment should make the Minigame rules more intuitive for the user.