Rename 'statusOnLogin' to 'showStatusOnLogin'
Renamed the variable 'statusOnLogin' to 'showStatusOnLogin' in the PlayerJoinListener class, PickpocketConfiguration class, and in the config.yml file to make its purpose more clear and improve code readability. Despite only being a small naming change, this makes the variable's purpose of controlling the display of status messages on login much more apparent.