Increase minigame clicks requirement, remove timer
Increased the number of required correct clicks in the minigame from 3 to 5. This was done to provide a more challenging user experience. Removed the timer and associated variables and functions from the minigame. This results in a more flexible and relaxed gameplay, as players were previously feeling too rushed and found the game difficult to enjoy.