Remove rummage timer from RummageInventory.kt
The rummage timer within RummageInventory.kt and associated methods were removed to improve the efficiency of the PickpocketPlugin. This change was necessary as the timer was introducing unnecessary complexity and delays. Now, the rummaging process is controlled by a predefined set of conditions, making the pickpocketing process more streamlined and straightforward.