Make required clicks to pickpocket configurable
A new configuration option "requiredClicksToPickpocket" was added to define the required number of clicks to successfully pickpocket. This provides better control over the difficulty of the pickpocketing feature, allowing it to be adjusted based on game balance needs. The hard-coded value "requiredClicks" used previously in Minigame.kt has been replaced with this new config variable.