Play sound for victim when user mis-clicks during minigame
A new sound has been added to play for the user being pickpocketed if the person attempting to pickpocket them mis-clicks during the minigame. Previously, the user being pickpocketed would only hear a sound if the person pickpocketing them clicked on the correct item. The purpose of this change is to improve the game experience by providing real-time audio feedback for the user being pickpocketed, alerting them to any unsuccessful pickpocket attempts.