Make 'toggle' and 'status' commands admin commands
This commit updates the plugin's command structure making the `toggle` and `status` commands in the admin section. The `PlayerJoinListener` class was deleted as it is no longer needed. The `PickpocketConfiguration` class was renamed to `Config` for simplicity and functionality involving toggling for individual users and status on login was removed. In line with these changes, option `pickpocketToggling` was replaced with `pickpocketingEnabled` in `config.yml` and associated methods were updated in various classes. Also, permission nodes for commands in `plugin.yml` were updated along with the changes in its command structure and commands information messages in `MainCommand.kt` was adjusted accordingly. This commit aims to address the issue of unnecessary complexity and make the plugin admin-focused.