Merged Victim and Thief profiles into a generic Profile
The ThiefProfile and VictimProfile along with ProfileType were removed and replaced with a generic Profile class. This change was made to simplify the profile handling as there was too much redundant code between the two Profile types. All references to Victim and Thief profiles were updated accordingly. Commands and Message configurations were also streamlined to adapt to the new structure.