Merge thief and victim profile types into one (#146)
* Fix incorrectly configured permissions

The permissions were previously misconfigured for certain commands within the 'pickpocket' plugin, causing unnecessary restrictions for users. This change corrects the issue, revising command permissions to properly reflect their original intent. Additionally, this update simplifies the command structure for profile manipulation, making it more user-friendly. Our aim with this commit is to ensure the plugin operates as intended and its features can be accessed & utilized with ease by users.

* Merged Victim and Thief profiles into a generic Profile

The ThiefProfile and VictimProfile along with ProfileType were removed and replaced with a generic Profile class. This change was made to simplify the profile handling as there was too much redundant code between the two Profile types. All references to Victim and Thief profiles were updated accordingly. Commands and Message configurations were also streamlined to adapt to the new structure.