Remove message displayed when a player moves while being pick-pocketed (#147)
* Fix incorrectly configured permissions

The permissions were previously misconfigured for certain commands within the 'pickpocket' plugin, causing unnecessary restrictions for users. This change corrects the issue, revising command permissions to properly reflect their original intent. Additionally, this update simplifies the command structure for profile manipulation, making it more user-friendly. Our aim with this commit is to ensure the plugin operates as intended and its features can be accessed & utilized with ease by users.

* Merged Victim and Thief profiles into a generic Profile

The ThiefProfile and VictimProfile along with ProfileType were removed and replaced with a generic Profile class. This change was made to simplify the profile handling as there was too much redundant code between the two Profile types. All references to Victim and Thief profiles were updated accordingly. Commands and Message configurations were also streamlined to adapt to the new structure.

* Remove hardcoded Profile type in Profile class

The hardcoded ProfileType.THIEF has been removed from Profile class to make it more flexible for assigning different profile types. This change helps promote code reusability and prepares for the introduction of new profile types.

* Changed "pickpocketDisabledOtherMessage" to "pickpocketDisabledMessage" in PlayerInteractListener.

This was done as the previous reference was to a non-existent message in MessageConfiguration.kt and messages.yml thus causing an error. Updated both the files to sync the relevant disabling messages preventing any possible missing message errors in the future.

* Removed message sent when a player moves while being pick-pocketed

This commit removes the redundant "pickpocketOnMoveWarningOther" message. This message was being sent to the last predator whenever a player moved, creating unnecessary notification. Simplifying this interaction enhances the user experience.